prescription_gen/requirements.txt
jinja2>=3.1
indic-transliteration>=2.3
fpdf2>=2.7.0
python-dotenv>=1.0
google-generativeai>=0.8.0
certifi>=2024.0.0
